Extract TrackerSizeTierIcon to shared component
Reuse the size tier ring icon in the settings panel size selector,
replacing the generic Minimize2/Maximize2 lucide icons. Also fix
temperature parsing to preserve decimal precision until display time.
